+++ CHANGES.md
@@ -4,6 +4,7 @@
   * Added `ThreadEmailService` for simple async sending of emails
   * Added basic logging with privacy configuration to mail class
   * Restructured documentation
+  * Restructured unit-tests
 
 * *2.0.0* (2024-04-11)
   * Dropped Django 3.2 & 4.1 support (via `ambient-package-update`)
